requirements.txt
discord.py==2.3.2
pandas
scikit-learn
xgboost
joblib
requests
